Sigmoid Added
Sigmoid and dSigmoid have been added to Tensor. Next steps will be figuring out integration into the neural network side of the library. I am worried that a traditional neural network doesn't apply to tensors in the way that I think it does. This might cause an early jump to convolutional NN.